+++ src/modules/voxel/Palette.h
@@ -21,6 +21,8 @@ class Palette {
 private:
 	bool load(const image::ImagePtr &img);
 	bool _dirty = false;
+	bool _needsSave = false;
+	core::String _paletteFilename;
 public:
 	PaletteColorArray colors {};
 	PaletteColorArray glowColors {};
@@ -31,14 +33,33 @@ public:
 		return colorCount;
 	}
 	bool load(const char *name);
-	bool save(const char *name);
+	bool save(const char *name = nullptr);
 	bool load(const uint8_t *rgbaBuf, size_t bufsize);
 
 	bool minecraft();
 	bool magicaVoxel();
 
-	bool isDirty() const;
-	void markClean();
+	void markDirty() {
+		_dirty = true;
+	}
+
+	bool isDirty() const {
+		return _dirty;
+	}
+
+	void markClean() {
+		_dirty = false;
+	}
+
+	void markSave() {
+		_needsSave = true;
+	}
+	bool needsSave() const {
+		return _needsSave;
+	}
+	void markSaved() {
+		_needsSave = false;
+	}
 
 	bool hasGlow(uint8_t idx) const;
 	void removeGlow(uint8_t idx);